chore: Delete app/scripts/vendor directory
In chapter 4 we started to wire in 3rd party components with
bower. And on page 74 of version 7 of the book we deleted
reference to the app/scripts/vendor directory from our index.html.
So it would be logical if we delete this directory. So in
chapter 5 it will be missing.